hub/requirements.txt
firebase-admin>=6.0.0
kivy>=2.3.0
sense-hat>=2.3.0
python-dotenv>=1.0.0
evdev>=1.7.0
requests>=2.31.0
